+++ lib/Elastica/Query/Builder.php
@@ -5,13 +5,15 @@ use Elastica\Exception\InvalidException;
 use Elastica\Exception\JSONParseException;
 use Elastica\JSON;
 
+trigger_error('This builder is deprecated and will be removed in Elastica 4.0. Use new Elastica\QueryBuilder instead.', E_USER_DEPRECATED);
+
 /**
  * Query Builder.
  *
  * @author Chris Gedrim <chris@gedr.im>
  *
  * @link http://www.elastic.co/
- * @deprecated This builder is deprecated and will be removed. Use new Elastica\QueryBuilder instead.
+ * @deprecated This builder is deprecated and will be removed in Elastica 4.0. Use new Elastica\QueryBuilder instead.
  **/
 class Builder extends AbstractQuery
 {