+++ lib/Elastica/Query/MoreLikeThis.php
@@ -1,6 +1,7 @@
 <?php
 namespace Elastica\Query;
 use Elastica\Document;
+use Elastica\Exception\DeprecatedException;
 
 /**
  * More Like This query.
@@ -28,13 +29,13 @@ class MoreLikeThis extends AbstractQuery
      *
      * @param array $ids Document ids
      *
-     * @deprecated Option "ids" deprecated as of ES 2.0.0-beta1 Use "like" instead.
+     * @deprecated Option "ids" deprecated as of ES 2.0.0-beta1 and will be removed in Elastica 4.0. Use "like" instead.
 
      * @return \Elastica\Query\MoreLikeThis Current object
      */
     public function setIds(array $ids)
     {
-        return $this->setParam('ids', $ids);
+        throw new DeprecatedException('Option "ids" deprecated as of ES 2.0.0-beta1 and will be removed in Elastica 4.0. Use "like" instead.');
     }
 
     /**
@@ -54,15 +55,13 @@ class MoreLikeThis extends AbstractQuery
      *
      * @param string $likeText
      *
-     * @deprecated Option "like_text" deprecated as of ES 2.0.0-beta1 Use "like" instead.
+     * @deprecated Option "like_text" deprecated as of ES 2.0.0-beta1 and will be removed at Elastica 4.0. Use "like" instead.
 
      * @return $this
      */
     public function setLikeText($likeText)
     {
-        $likeText = trim($likeText);
-
-        return $this->setParam('like_text', $likeText);
+        throw new DeprecatedException('Option "like_text" deprecated as of ES 2.0.0-beta1 and will be removed in Elastica 4.0. Use "like" instead.');
     }
 
     /**
@@ -96,11 +95,11 @@ class MoreLikeThis extends AbstractQuery
      *
      * @return $this
      *
-     * @deprecated Option "percent_terms_to_match" deprecated as of ES 1.5. Use "minimum_should_match" instead.
+     * @deprecated Option "percent_terms_to_match" deprecated as of ES 1.5 and will be removed in Elastica 4.0. Use "minimum_should_match" instead.
      */
     public function setPercentTermsToMatch($percentTermsToMatch)
     {
-        return $this->setParam('percent_terms_to_match', (float) $percentTermsToMatch);
+        throw new DeprecatedException('Option "percent_terms_to_match" deprecated as of ES 1.5 and will be removed in Elastica 4.0. Use "minimum_should_match" instead.');
     }
 
     /**