+++ lib/Elastica/QueryBuilder/DSL/Query.php
@@ -1,6 +1,7 @@
 <?php
 namespace Elastica\QueryBuilder\DSL;
 
+use Elastica\Exception\DeprecatedException;
 use Elastica\Exception\NotImplementedException;
 use Elastica\Filter\AbstractFilter;
 use Elastica\Query\AbstractQuery;
@@ -340,11 +341,11 @@ class Query implements DSL
      * more_like_this_field query.
      *
      * @link http://www.elastic.co/guide/en/elasticsearch/reference/1.4/query-dsl-mlt-field-query.html
-     * @deprecated More Like This Field query is deprecated as of ES 1.4 and will be removed in ES 2.0
+     * @deprecated More Like This Field query is deprecated as of ES 1.4 and will be removed in ES 2.0. Use MoreLikeThis query instead. This method will be removed in Elastica 4.0
      */
     public function more_like_this_field()
     {
-        throw new NotImplementedException();
+        throw new DeprecatedException('More Like This Field query is deprecated as of ES 1.4 and will be removed in ES 2.0. Use MoreLikeThis query instead. This method will be removed in Elastica 4.0');
     }
 
     /**